Break store out of DefaultPropertyCursor
